message: move some functions to message package
Closes #53
