+++ packages/server/src/utils.ts
@@ -77,26 +77,3 @@ export function verifyByKeyType(
       throw new Error('Unreachable Code');
   }
 }
-
-export function propertyObject(properties: Property[]): { [key: string]: string } {
-  return properties.reduce(
-    (dict: { [key: string]: string }, property) => {
-      if (property.type !== PropertyType.Hash) {
-        return {
-          ...dict,
-          [property.key]: property.value,
-        };
-      }
-      return dict;
-    },
-    {},
-  );
-}
-
-export function objectToProperty(object: { [key: string]: string }): Property[] {
-  return Object.keys(object).map(key => ({
-    type: PropertyType.Raw,
-    key,
-    value: object[key]
-  }))
-}
